+++ src/RcsbFvExamples/CompositeTrack.ts
@@ -45,7 +45,7 @@ const rowConfigData: RcsbFvRowConfigInterface[] = [
                 displayId:"compositeSeqeunce",
                 displayData: [{
                     begin:50,
-                    value:"TCLLDELDTAGQEEYSAMRDQYMRTSEGFLC"
+                    label:"TCLLDELDTAGQEEYSAMRDQYMRTSEGFLC"
                 }]
             }
         ]
@@ -84,7 +84,7 @@ const sequenceConfigData: RcsbFvRowConfigInterface[] = [
         trackData: [
             {
                 begin: 1,
-                value: sequence
+                label: sequence
             }
         ]
     }];